refactor(cli): extract Kafka request parsers into dedicated module
Move the parsing/validation surface of `/api/kafka/endpoint` out of
`packages/cli/src/daemon/routes/kafka.ts` into a new module at
`packages/cli/src/daemon/parsers/kafka-request.ts`. The route file now
only carries route-handler logic, error translation, and HTTP glue.

Moved verbatim (no rename, no signature change):
- `isNonEmptyString`
- `VALID_PROTOCOLS`, `VALID_SASL_MECHANISMS`
- `parseSecurityProtocol`, `parseSasl`, `parseSsl`, `shouldProbe`
- `KafkaEndpointRequestBody`

The route imports the helpers back from the new module; the parser
test (`packages/cli/test/kafka-route-parsers.test.ts`) updates its
import path to match. No behavior change.
